+++ pkgs/artisanal/lib/plugins.dart
@@ -38,6 +38,8 @@ export 'src/plugins/remote_surface_transport.dart'
 export 'src/plugins/remote_surface_channel.dart' show RemotePluginJsonChannel;
 export 'src/plugins/remote_surface_controller.dart'
     show RemotePluginSurfaceController;
+export 'src/plugins/remote_surface_drawable.dart'
+    show RemotePluginSurfaceDrawable;
 export 'src/plugins/remote_surface_guest_session.dart'
     show RemotePluginGuestSession;
 export 'src/plugins/remote_surface_session.dart' show RemotePluginSession;